chore: Update composer.json, seeders, routes, and factories
This commit updates the composer.json file by removing the "spatie/laravel-medialibrary" dependency. It also adds a new seeder class, AppUserSeeder, which creates 10 instances of the AppUser model. In the routes/api.php file, two new routes are added: one for retrieving all users and another for finding a user by email. Additionally, the DatabaseSeeder class is modified to create a new user with the name "User Admin" and email "kisama972@gmail.com". Finally, the SlideResource form schema is updated to make the "image" field required.